Image compression and downscaling; remove background added
The output .htm files can be quite big (10+MB) due to the embedded images. I added an option in Options to compress and scale down images to the size they are displayed in. The compression level is changeable by the user.

I also added an option to remove background image from the output, as often this would be the reason for a huge .htm output file.

Last I added a few comments to help with following the flow of the code, which proved quite difficult when I was first getting into it.